Commit changes from tarball openwebmail-1.80.tar.gz
git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@79 9f51437f-89f6-0310-a5ff-ff9a3806d2ee

+++ src/cgi-bin/openwebmail/uty/mktgz.sh
@@ -39,6 +39,7 @@ cp /dev/null cgi-bin/openwebmail/etc/calendar.book
 #cp /dev/null cgi-bin/openwebmail/etc/filter.book
 patch cgi-bin/openwebmail/etc/openwebmail.conf < /usr/local/www/cgi-bin/openwebmail/uty/openwebmail.conf.diff
 rm cgi-bin/openwebmail/etc/openwebmail.conf.orig
+rm cgi-bin/openwebmail/etc/openwebmail.conf.rej
 
 tar -zcBpf /tmp/openwebmail-current.tgz *
 
@@ -46,32 +47,37 @@ tar -zcBpf /tmp/openwebmail-current.tgz *
 ##########################################################
 
 oldtgz=`ls -tr /usr/local/www/data/openwebmail/download/openwebmail-?.*tgz|tail -1`
-oldrelease=`echo $oldtgz|sed -e 's/.*openwebmail-//'|sed -e 's/\.tgz.*//'`
+if [ ! -z "$oldtgz" ]; then
+   oldrelease=`echo $oldtgz|sed -e 's/.*openwebmail-//'|sed -e 's/\.tgz.*//'`
 
-echo make openwebmail-current-$oldrelease.diff.gz
+   echo make openwebmail-current-$oldrelease.diff.gz
 
-cd /tmp
-mv openwebmail-current www
+   cd /tmp
+   mv openwebmail-current www
 
-mkdir www.orig
-cd www.orig
-tar -zxBpf $oldtgz
-cd ..
+   mkdir www.orig
+   cd www.orig
+   tar -zxBpf $oldtgz
+   cd ..
 
-rm -Rf \
-www.orig/data/openwebmail/images \
-www.orig/data/openwebmail/*.wav \
-www.orig/data/openwebmail/*.au \
-www.orig/data/openwebmail/index.html \
-www/data/openwebmail/images \
-www/data/openwebmail/*.wav \
-www/data/openwebmail/*.au \
-www/data/openwebmail/index.html
+   rm -Rf \
+   www.orig/data/openwebmail/images \
+   www.orig/data/openwebmail/*.wav \
+   www.orig/data/openwebmail/*.au \
+   www.orig/data/openwebmail/index.html \
+   www/data/openwebmail/images \
+   www/data/openwebmail/*.wav \
+   www/data/openwebmail/*.au \
+   www/data/openwebmail/index.html
 
-diff -ruN www.orig www|grep -v '^Binary files '|gzip -9>openwebmail-current-$oldrelease.diff.gz
+   diff -ruN www.orig www|grep -v '^Binary files '|gzip -9>openwebmail-current-$oldrelease.diff.gz
+   cp www.orig/cgi-bin/openwebmail/etc/openwebmail.conf.default openwebmail.conf.default.orig
+   cp www/cgi-bin/openwebmail/etc/openwebmail.conf.default      openwebmail.conf.default
+   diff -ruN openwebmail.conf.default.orig openwebmail.conf.default >openwebmail.conf.default-current-$oldrelease.diff
 
-rm -Rf www.orig
-mv www openwebmail-current
+   rm -Rf www.orig openwebmail.conf.default.orig openwebmail.conf.default
+   mv www openwebmail-current
+fi
 
 ##########################################################
 
@@ -79,8 +85,8 @@ echo cp new stuff to download
 
 cp /usr/local/www/data/openwebmail/doc/*.txt /usr/local/www/data/openwebmail/download/doc/
 
-rm /usr/local/www/data/openwebmail/download/openwebmail-current*gz
-for f in openwebmail-current.tgz openwebmail-current-$oldrelease.diff.gz; do
+rm /usr/local/www/data/openwebmail/download/openwebmail*current*gz 
+for f in openwebmail-current.tgz openwebmail-current-$oldrelease.diff.gz openwebmail.conf.default-current-$oldrelease.diff; do
    mv /tmp/$f /usr/local/www/data/openwebmail/download/
    chmod 644 /usr/local/www/data/openwebmail/download/$f
 done
